Extract shared formatUtcTimestamp helper
